refactor: align finals naming with final-exams convention
Component dir finals/ -> final-exams/ (matches /final-exams route and
src/lib/final-exams); fold final-exams-calendar.test.ts into
term-calendar.test.ts, which it was actually testing.
